fast foreign key integrity check for integer keys
Added function "check_foreign_keys". Disables regular SQL FK checking
as it slows down bulk uploads of huge tables considerably.
When FKs are added after the insert, it still
is a painfully slow process when millions of unique Key values
are involved.

Only works for integer columns.
As for the referenced column, following conditions are required:
    a) it is a Primary Key (ensuring that it's values are unique)
    b) all distinct values between MIN and MAX are present
       (MAX value of Primary Key = number of rows in parent table)